requirements.txt
numpy
opencv-python
pyyaml
matplotlib
